requirements.txt
# requirements.txt
psycopg2-binary==2.9.9
structlog==24.4.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1